bin/make-fai-nfsroot: do not divert update-initramfs by default any longer
The FAI generated /etc/kernel-img.conf avoids generation
of the initrd by the kernel by default. The divertion
mechanism from make-fai-nfsroot avoids registration of
the kernel for initramfs-tools and therefor generating
the initrd will not work anymore without specifying the
exact kernel version (instead of just using update-initramfs
with "-k all"). If we avoid the divertion by default
the initrd might be build twice during the setup procedure
but we do not break it, so this should be a good tradeoff.

The -U option ("do not divert") doesn't change the behaviour
anymore but is still available for backward compability
reasons.

If someone still wants to explicitely divert the update-initramfs
script we could introduce a new variable to support that, but
for now lets KISS.

Acked-By: Thomas Lange <lange@debian.org>
Signed-off-by: Michael Prokop <mika@grml.org>

git-svn-id: svn+ssh://svn.debian.org/svn/fai@6043 ba5ec265-b0fb-0310-8e1a-cf9e4c2b1591