fai: re-add support for hostname=... bootoption when using live-config
live-config provides hostname handling through a separate script
and $HOSTNAME of bootoption hostname=... isn't available in the
environment any longer.

The argument of the bootoption HOSTNAME=... (uppercase) would be
available in the environment but as we want to be backwards
compatible and continue supporting the hostname=... bootoption
we've to use live-config accordingly if it's available.

Thanks for the bugreport and the initial patch to Karl Wallner.

Thanks: Karl Wallner <fai@karlwallner.de>
Signed-off-by: Michael Prokop <mika@debian.org>



git-svn-id: svn+ssh://svn.debian.org/svn/fai@6124 ba5ec265-b0fb-0310-8e1a-cf9e4c2b1591