add initramfs-tools, workaround for fai-mirror problem
When creating a mirror using fai-mirror, the package linux-image-amd64
will not be added to the mirror because of aptitude's dependency
resolver. This worked with aptitude from lenny, but the squeeze
version will refuse to install/download the kernel package.
Using apt-get will also work, but currently adding initramfs-tools
will fix this issue.


git-svn-id: svn+ssh://svn.debian.org/svn/fai@6455 ba5ec265-b0fb-0310-8e1a-cf9e4c2b1591