call genisoimage with -R instead of -r
Quote from a mail:
According to the genisoimage man page the "-r" does what normally would be
consider the sane thing and sets all directories to 555. This makes sense as
most people will want the cdfs to be readable by normal users. The problem is
that cfagent doesn't like it's /var/lib directories to be readable by others
and will not run.
The "-R" tells genisoimage to copy the perms as specified in the source
directory. With this option cfagent runs correctly and all is well.

Me: Another approach would be that the permissions for
/var/lib/cfengine2 are fixed during runtime.


git-svn-id: svn+ssh://svn.debian.org/svn/fai@6522 ba5ec265-b0fb-0310-8e1a-cf9e4c2b1591