clean up the name space: functions in setup-storage
remove the "&FAI::" from all the functions used by setup-storage.

To be able to do that, the external functions are exported from
the FAI.pm. The renaming to FAI.pm is necessary because the
namespace and the file name of the module have to match.

We can rename the File later on to something more meaningfull
again.

I also introduced a HELP_MESSAGE that is called if setup-storage
is called with --help.